TTScore now 32bit, like Score

+++ src/types.h
@@ -46,7 +46,7 @@ typedef u64 Hash;
 typedef u64     TTMove;
 typedef s32     MoveScore;
 typedef s16     Score;
-typedef s16     TTScore;
+typedef s32     TTScore;
 typedef u8      Square;
 typedef u8      Piece;
